+++ CHANGELOG.md
@@ -12,6 +12,7 @@ with the current date and the next changes should go under a **[Next]** header.
 * Add confirmation prompt for cancelling and deleting student questions. ([@muakasan](https://github.com/muakasan) in [#71](https://github.com/illinois/queue/pull/71))
 * Hide new question panel for active course staff. ([@genevievehelsel](https://github.com/genevievehelsel) in [#61](https://github.com/illinois/queue/pull/61))
 * Updated README with information about the queue as a service for users at Illinois. ([@wadefagen](https://github.com/wadefagen) in [#73](https://github.com/illinois/queue/pull/73))
+* Fix the database schema for `questions.preparedness`. ([@nwalters512](https://github.com/nwalters512) in [#75](https://github.com/illinois/queue/pull/75))
 
 ## 13 March 2018